Shape-agnostic circuit-finder ((a)+(b) upgrade): cave output-carrier is DISTRIBUTED + MLP-heavy
Set-aware finder (fixed items + ATP over every head/MLP + joint size-sweep confirm +
doubt-classed-set reference; the per-component confirm was set-blind, ATP under-ranks
redundant sets). 9b base, n=27. Verdict DISTRIBUTED: conc_frac(top-15)=0.29, joint
restore by K=1/3/5/10/15 = 0.00/0.00/0.04/0.36/0.79 (needs ~15, no small set);
ATP-top components mostly MLPs (MLP24/31/23 confirm 0.35/0.37/0.13). The ATP-top
DOUBT-classed heads (3) jointly output-patch to ~0 (0.0025) -- doubt heads are NOT the
output-carriers.

Tension with the head-set finding (doubt attention-KO restored 0.59), unresolved: (i)
different intervention -- attention-KO removes the READ (input gating), output-patch
removes the WRITE; (ii) ATP under-ranks redundant sets so the ATP-doubt-3 != the
span-doubt-5 (not a direct comparison). Honest status: cave output-carrier is
distributed/MLP, NOT a concentrated doubt-head WRITE circuit; the doubt-READ gate is a
separate input effect. Doubt circuit NOT proven as a write circuit.

Resolution next: output-patch the head-set's span-doubt-5 jointly (direct), + DLA-link
cross-check (pending, ssh-abort 255). Both DLA fns fixed (bf16->float; device align).

+++ controls/cave_doubt_writes_cavedir.py
@@ -198,7 +198,7 @@ def dla_logit_diff(vec, ln_final, W_U, b_U, cap, wstar_id, c_id):
     (tensors + ids -> float). `ln_final` is a callable (the model's final LayerNorm) or a pure function in
     the selftest."""
     h = ln_final(vec.float().unsqueeze(0))[0]            # [d_model]; final LN exactly as applied at readout
-    h = h.to(W_U.device)                                 # head writes are cpu; W_U is on-device -> align
+    h = h.to(W_U.device).float()                         # align device AND dtype (ln_final returns bf16; W_U.float() is f32)
     logits = h @ W_U.float() + (b_U.float() if b_U is not None else 0.0)   # [d_vocab]
     logits = softcap(logits, cap)
     return float(logits[wstar_id] - logits[c_id])
